requirements-test.txt
flake8
mongomock
pytest
pytest-cov
requests-mock
